Email Analysis Result Dialog/Spionner HTML and functions moved into separate folder. Plugin code cleanup done
